Add identity depository and remove mango code
