+++ README.md
@@ -47,40 +47,6 @@ import { Connection } from '@solana/web3.js';
 const connection = new Connection('https://api.mainnet-beta.solana.com');
 ```
 
-### Create and initialize the Mango object
-
-```javascript
-import { createAndInitializeMango } from '@uxd-protocol/uxd-client';
-
-const mango = createAndInitializeMango(
-  connection,
-  'mainnet' // 'devnet'
-);
-```
-
-### Instantiate a MangoDepository Object for a given UXD collateral, ie WSOL
-
-```javascript
-import {
-  WSOL,
-  SOL_DECIMALS,
-  USDC,
-  USDC_DECIMALS,
-  MangoDepository,
-} from '@uxd-protocol/uxd-client';
-const mainnetProgramId = new PublicKey('UXD8m9cvwk4RcSxnX2HZ9VudQCEeDH6fRnB4CAP57Dr');
-
-const depository = new MangoDepository(
-  WSOL, 
-  'SOL',
-  SOL_DECIMALS,
-  USDC, // Use mainnet mint, must be matching the program used (see USDC_DEVNET)
-  'USDC',
-  USDC_DECIMALS,
-  mainnetProgramId
-);
-```
-
 ### Instantiate the UXD Client
 
 ```javascript
@@ -89,139 +55,4 @@ import { UXDClient } from '@uxd-protocol/uxd-client';
 const mainnetProgramId = new PublicKey('UXD8m9cvwk4RcSxnX2HZ9VudQCEeDH6fRnB4CAP57Dr');
 
 const client = new UXDClient(mainnetProgramId);
-```
-
-### Mint 1 SOL worth of UXD on the SOL Mango Depository
-
-```javascript
-import { Transaction } from '@solana/web3.js';
-
-const transaction = new Transaction();
-const mintWithMangoDepositoryIx =
-  await client.createMintWithMangoDepositoryInstruction(
-    1, // UI amount of collateral to use to mint UXD
-    5, // 0.5% slippage in points
-    controller,
-    depository, // matching Mango Depository for the collateral
-    mango,
-    user // PublicKey of recipient
-    {}, // ConfirmOptions, leave empty for defaults
-    payer, // PublicKey of payer, optional, defaults to user
-  );
-
-transaction.add(mintWithMangoDepositoryIx);
-
-// sign, send & confirm transaction
-```
-
-### Redeem UXD on the SOL Mango Depository
-
-```javascript
-import { Transaction } from '@solana/web3.js';
-
-const transaction = new Transaction();
-const redeemFromMangoDepositoryIx =
-  await client.createRedeemFromMangoDepositoryInstruction(
-    10, // UI amount of UXD to redeem (returned as the provided depository Collateral, here SOL))
-    5, // 0.5% slippage in points
-    controller,
-    depository, // matching Mango Depository for the collateral
-    mango,
-    user // PublicKey of recipient
-    {}, // ConfirmOptions, leave empty for defaults
-    payer, // PublicKey of payer, optional, defaults to user
-  );
-
-transaction.add(redeemFromMangoDepositoryIx)
-
-// sign, send & confirm transaction
-```
-
-### Rebalance an UXD Mango Depository
-
-This can be used as a Swap. Depending of the polarity of the PnL of the depository, this instruction
-will switch Quote->Collateral or Collateral->Quote.
-
-For instance, when the PnL is Positive (Collateral price went up since minting), user can provide USDC.
-By doing so, UXDProgram will offset the PnL of an equal amount, close equivalent short position, and return the backing collateral amount.
-
-It's a 4bps + slippage + spread swap (as we use the Perp price for all operations).
-
-You can find detailed information about the current state of depositories on the [UXD Backoffice](https://backoffice.uxd.fi).
-
-```javascript
-import { Transaction } from '@solana/web3.js';
-
-const unrealizedPnl = await depository.getUnrealizedPnl(mango, TXN_OPTS);
-const polarity = unrealizedPnl > 0 ? PnLPolarity.Positive : PnLPolarity.Negative;
-
-const transaction = new Transaction();
-const rebalanceLiteMangoDepositoryIx =
-  await client.createRebalanceMangoDepositoryLiteInstruction(
-    10, // UI amount in Quote to rebalance (Quote on mango is USDC)
-    5, // 0.5% slippage in points
-    PnLPolarity.positive, // The polarity of the PnL, the expected direction of the swap (can be enforce program side but here to be explicit)
-    controller,
-    depository, // matching Mango Depository for the collateral
-    mango,
-    user // PublicKey of recipient
-    {}, // ConfirmOptions, leave empty for defaults
-    payer, // PublicKey of payer, optional, defaults to user
-  );
-
-transaction.add(redeemFromMangoDepositoryIx)
-
-// sign, send & confirm transaction
-```
-
-### Retrieve the price of the depository Perp (mid price)
-
-```javascript
-const perpPrice = await depository.getCollateralPerpPriceUI(mango);
-```
-
-### Retrieve the price impact and minting/redeeming estimates
-
-To determine the estimates about a minting operation :
-
-```javascript
-// User wants to mint `collateralQuantity` (Collateral -> UXD)
-const collateralQuantity = 5; // 5 SOL
-const mintingPriceImpact = await depository.getMintingPriceImpact(
-  collateralQuantity,
-  mango
-);
-console.log("Minting price impact", mintingPriceImpact);
-
-const mintingEstimates = depository.getMintingEstimates(
-  collateralQuantity,
-  perpPrice,
-  mintingPriceImpact,
-  mango,
-);
-console.log("Will mint", mintingEstimates.yield, "UXD");
-console.log("(fees:", mintingEstimates.fees, ", slippage:", mintingEstimates.slippage, ")");
-```
-
-and for redeeming :
-
-```javascript
-// User wants to redeem `redeemableQuantity` of UXD (UXD -> Collateral)
-const redeemableQuantity = 100; // 100 UXD
-const redeemingPriceImpact = await depository.getRedeemingPriceImpact(
-  redeemableQuantity,
-  perpPrice,
-  mango
-);
-console.log("Redeeming price impact", redeemingPriceImpact);
-
-const redeemingEstimates = depository.getRedeemingEstimates(
-  redeemableQuantity,
-  perpPrice,
-  redeemingPriceImpact,
-  mango,
-);
-console.log("Will redeem", redeemingEstimates.yield, "Collateral");
-console.log("(fees:", redeemingEstimates.fees, ", slippage:", redeemingEstimates.slippage, ")");
-
 ```
