+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@uxd-protocol/uxd-client",
-  "version": "5.0.0",
+  "version": "5.1.0",
   "description": "JavaScript Client for the UXD Solana Program",
   "keywords": [
     "solana",
@@ -38,7 +38,6 @@
     "validate": "npm run typecheck && npm run test && npm run lint && npm run format-check"
   },
   "dependencies": {
-    "@blockworks-foundation/mango-client": "3.4.7",
     "@project-serum/anchor": "0.24.2",
     "@solana/spl-token": "0.1.8",
     "@solana/web3.js": "1.48.0",
